+++ IMoBroPlugin.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
+using MoBro.Plugin.SDK.Services;
 
 namespace MoBro.Plugin.SDK;