+++ Models/Actions/IAction.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using MoBro.Plugin.SDK.Models.Settings;
+using MoBro.Plugin.SDK.Services;
 
 namespace MoBro.Plugin.SDK.Models.Actions;